requirements.txt
opencv-python>=4.11.0.86
numpy>=2.0.2
Pillow>=11.1.0
flask==3.1.0
tensorflow>=2.18.0
werkzeug>=3.1.3
